requirements.txt
requests==2.23.0
tempo-api-python-client==0.4.1
PyYAML==5.3.1
